Pass size_t to span functions: chrome/
In preparation for upgrading various span args from size_t to
StrictNumeric<size_t>, this fixes type mismatches that would cause
compile errors with that change.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020621
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382641}

+++ chrome/browser/enterprise/connectors/analysis/analysis_service_settings_unittest.cc
@@ -730,7 +730,7 @@ class AnalysisServiceSettingsTest : public testing::TestWithParam<TestParam> {
       GURL regionalized_url =
           GURL(GetServiceProviderConfig()
                    ->at("google")
-                   .analysis->region_urls[static_cast<int>(data_region())]);
+                   .analysis->region_urls[static_cast<size_t>(data_region())]);
       absl::get<CloudAnalysisSettings>(
           GetParam().expected_settings->cloud_or_local_settings)
           .analysis_url = regionalized_url;
